+++ processor/src/test/fixtures/expected/com/example/injector/eager/EagerInjectableViaIncludesModel_Sting_MyFragment1.java
@@ -1,7 +1,0 @@
-package com.example.injector.eager;
-
-import javax.annotation.Generated;
-
-@Generated("sting.processor.StingProcessor")
-public final class EagerInjectableViaIncludesModel_Sting_MyFragment1 implements EagerInjectableViaIncludesModel.MyFragment1 {
-}