+++ processor/src/test/fixtures/expected/com/example/injector/eager/fragment/Sting_EagerInjectableViaIncludesModel.java
@@ -1,4 +1,4 @@
-package com.example.injector.eager;
+package com.example.injector.eager.fragment;
 
 import java.util.Objects;
 import javax.annotation.Generated;
@@ -7,13 +7,9 @@ import javax.annotation.Nonnull;
 @Generated("sting.processor.StingProcessor")
 final class Sting_EagerInjectableViaIncludesModel implements EagerInjectableViaIncludesModel {
   @Nonnull
-  private final Object node1;
-
-  @Nonnull
-  private final Object node2;
+  private final MyModel3 node1;
 
   Sting_EagerInjectableViaIncludesModel() {
-    node1 = Objects.requireNonNull( EagerInjectableViaIncludesModel_Sting_MyModel3.create() );
-    node2 = Objects.requireNonNull( EagerInjectableViaIncludesModel_Sting_MyModel1.create() );
+    node1 = Objects.requireNonNull( Sting_MyModel3.create() );
   }
 }