+++ processor/src/test/fixtures/expected/com/example/injector/eager/fragment/Sting_MyModel1.java
@@ -1,0 +1,15 @@
+package com.example.injector.eager.fragment;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+
+@Generated("sting.processor.StingProcessor")
+public final class Sting_MyModel1 {
+  private Sting_MyModel1() {
+  }
+
+  @Nonnull
+  public static MyModel1 create() {
+    return new MyModel1();
+  }
+}