+++ processor/src/test/fixtures/input/com/example/injector/eager/fragment/EagerInjectableViaIncludesModel.java
@@ -1,0 +1,8 @@
+package com.example.injector.eager.fragment;
+
+import sting.Injector;
+
+@Injector( includes = MyFragment1.class )
+interface EagerInjectableViaIncludesModel
+{
+}