+++ processor/src/test/fixtures/input/com/example/injector/eager/fragment/MyFragment1.java
@@ -1,0 +1,8 @@
+package com.example.injector.eager.fragment;
+
+import sting.Fragment;
+
+@Fragment( includes = { MyFragment2.class, MyModel2.class } )
+public interface MyFragment1
+{
+}