+++ processor/src/test/fixtures/input/com/example/injector/eager/fragment/MyFragment2.java
@@ -1,0 +1,8 @@
+package com.example.injector.eager.fragment;
+
+import sting.Fragment;
+
+@Fragment( includes = MyModel3.class )
+public interface MyFragment2
+{
+}