+++ processor/src/test/fixtures/input/com/example/injector/eager/fragment/MyModel1.java
@@ -1,0 +1,10 @@
+package com.example.injector.eager.fragment;
+
+import sting.Eager;
+import sting.Injectable;
+
+@Eager
+@Injectable
+public class MyModel1
+{
+}