+++ processor/src/test/fixtures/input/com/example/injector/eager/fragment/MyModel2.java
@@ -1,0 +1,8 @@
+package com.example.injector.eager.fragment;
+
+import sting.Injectable;
+
+@Injectable
+public class MyModel2
+{
+}